Start of reading album from a directory, including finding image files
