+++ Cargo.toml
@@ -10,6 +10,7 @@ edition = "2024"
 [dependencies]
 anyhow = "^1.0"
 clap = { version = "^4.5", features = ["derive"] }
+image = "^0.25.6"
 serde = { version = "^1.0", features = ["derive"] }
 serde_yml = "^0.0.12"
 thiserror = "^2.0"
